refactor: moved AsyncTaskRunner and IAsyncTask into AsyncTasks namespace
